% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadb2c_js.R
\name{aadb2c_js}
\alias{aadb2c_js}
\title{Generate some initial JS code to authenticate against AAD B2C}
\usage{
aadb2c_js(
  output_dir = ".",
  shiny_input = "email",
  tenant = "shortname",
  client_id = "guid",
  signin_policy = "B2C_1_signup_signin",
  scopes = c("email", "openid", "https://someresource/read")
)
}
\arguments{
\item{output_dir}{Where the JS file should be output}

\item{shiny_input}{The id of the input that should be given the user's email}

\item{tenant}{The AAD B2C tenant name (no onmicrosoft.com bit)}

\item{client_id}{The application ID of the AAD B2C app}

\item{signin_policy}{The User Flow / Policy to use to Sign In}

\item{scopes}{The scopes that should be requested, including a URI. Typically
the URI is a variant of the App API URI}
}
\value{
Configured JS file's location
}
\description{
Generate some initial JS code to authenticate against AAD B2C
}
\examples{
on.exit(aadb2c_js(tempdir()), unlink(tempdir()))
}
