% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadb2c_server.R
\name{aadb2c_server}
\alias{aadb2c_server}
\title{Generate some sample server code to view authentication results}
\usage{
aadb2c_server(shiny_input = "email")
}
\arguments{
\item{shiny_input}{Expected name of shiny input from JS}
}
\value{
Sample code to console and clipboard if available
}
\description{
Generate some sample server code to view authentication results
}
\examples{
aadb2c_server()
}
