% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadb2c_ui.R
\name{aadb2c_ui}
\alias{aadb2c_ui}
\title{Generate some sample UI code to perform authentication}
\usage{
aadb2c_ui()
}
\value{
Sample code to console and clipboard if available
}
\description{
Generate some sample UI code to perform authentication
}
\examples{
aadb2c_ui()
}
